# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/users/{username}": {
      "get": {
        "operationId": "UsersController_getByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicView"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [
          {
            "name": "x-client-type",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "x-client-type",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "x-client-type",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/folders/{folderId}/files": {
      "get": {
        "operationId": "FilesController_findAll",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/folders/{folderId}/files/presign": {
      "post": {
        "operationId": "FilesController_presign",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/folders/{folderId}/files/confirm": {
      "post": {
        "operationId": "FilesController_confirm",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/files/{id}": {
      "delete": {
        "operationId": "FilesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/feed": {
      "get": {
        "operationId": "PostsController_feed",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco devuelto por la página anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostResponseDto"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/posts/reorder": {
      "patch": {
        "operationId": "PostsController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderPostsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReorderResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{username}/posts": {
      "get": {
        "operationId": "PostsController_findByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco devuelto por la página anterior.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostResponseDto"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Micelio API",
    "description": "Auth, folders and file storage API for Micelio",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "FeedSettingsView": {
        "type": "object",
        "properties": {
          "layout": {
            "type": "string",
            "enum": [
              "GRID",
              "MASONRY"
            ]
          },
          "columns": {
            "type": "number"
          },
          "gap": {
            "type": "number"
          }
        },
        "required": [
          "layout",
          "columns",
          "gap"
        ]
      },
      "MeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "followersCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "viewerFollows": {
            "type": "boolean"
          },
          "followsViewer": {
            "type": "boolean"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsView"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "TEACHER",
              "ADMIN",
              "SUPPORT"
            ]
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "avatarUrl",
          "isPublic",
          "followersCount",
          "followingCount",
          "viewerFollows",
          "followsViewer",
          "email",
          "role"
        ]
      },
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "bio": {
            "type": "string",
            "example": "Ilustradora y animadora."
          },
          "isPublic": {
            "type": "boolean",
            "example": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsDto"
          }
        }
      },
...
      "UserPublicView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "followersCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "viewerFollows": {
            "type": "boolean"
          },
          "followsViewer": {
            "type": "boolean"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "feedSettings": {
            "$ref": "#/components/schemas/FeedSettingsView"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "avatarUrl",
          "isPublic",
          "followersCount",
          "followingCount",
          "viewerFollows",
          "followsViewer"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "S3cure-Password!"
          },
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "username": {
            "type": "string",
            "example": "ada.lovelace"
          },
          "cedula": {
            "type": "string",
            "example": "1020304050",
            "description": "Cédula colombiana (solo dígitos)"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "username",
          "cedula"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "S3cure-Password!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "FileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "folderId": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "TEXT"
            ]
          },
          "size": {
            "type": "number"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "url": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "folderId",
          "originalName",
          "mimeType",
          "type",
          "size",
          "width",
          "height",
          "url",
          "createdAt"
        ]
      },
      "PresignFileDto": {
        "type": "object",
        "properties": {
          "originalName": {
            "type": "string",
            "example": "boceto.png"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "size": {
            "type": "number",
            "example": 245678
          }
        },
        "required": [
          "originalName",
          "mimeType",
          "size"
        ]
      },
      "PresignResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string",
            "description": "URL firmada para un PUT directo del binario a S3."
          },
          "expiresIn": {
            "type": "number",
            "description": "Segundos de validez de uploadUrl."
          }
        },
        "required": [
          "key",
          "uploadUrl",
          "expiresIn"
        ]
      },
      "ConfirmFileDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "users/uuid/folders/uuid/uuid.png"
          },
          "originalName": {
            "type": "string",
            "example": "boceto.png"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "size": {
            "type": "number",
            "example": 245678
          },
          "width": {
            "type": "number",
            "description": "Ancho en píxeles, medido por el cliente. El binario nunca pasa por el backend, así que nadie más puede conocerlo. Opcional: sin él, el medio se dibuja cuadrado en el masonry.",
            "example": 1080
          },
          "height": {
            "type": "number",
            "description": "Alto en píxeles (ver `width`).",
            "example": 1350
          }
        },
        "required": [
          "key",
          "originalName",
          "mimeType",
          "size"
        ]
      },
...
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/UserPublicView"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "position": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMediaResponseDto"
            }
          },
          "viewerHasLiked": {
            "type": "boolean"
          },
          "viewerHasSaved": {
            "type": "boolean"
          },
          "likeCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "author",
          "description",
          "tags",
          "position",
          "createdAt",
          "media",
          "viewerHasLiked",
          "viewerHasSaved",
          "commentCount"
        ]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Bocetos del mural #proceso"
          },
          "tags": {
            "description": "Etiquetas explícitas. El servidor las normaliza y les suma los #hashtags de la descripción; si tras normalizar quedan más de 10, responde 400.",
            "example": [
              "ilustración",
              "proceso"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "media": {
            "description": "Medios en el orden del carrusel.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMediaInputDto"
            }
          }
        },
        "required": [
          "media"
        ]
      },
      "ReorderPostsDto": {
        "type": "object",
        "properties": {
          "orderedIds": {
            "description": "Todos los posts del usuario, en el orden deseado.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "orderedIds"
        ]
      },
      "ReorderResponseDto": {
        "type": "object",
        "properties": {
          "reordered": {
            "type": "boolean"
          }
        },
        "required": [
          "reordered"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "object",
            "example": "Versión final del mural #proceso"
          },
          "tags": {
            "example": [
              "mural",
              "proceso"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMediaInputDto"
            }
          }
        }
      },
...
    }
  }
}